Realized, we don't need image processing for teleop. So I put in a boolean to control whether we want just a simple camera so the driver can see what they're doing, or a more advance one for image processing.
This should give us the flexibility to go to plan B (no image
processing) very easily. As well, this might be faster than manually
sending a frame (the library code says it takes 50ms to manually
transfer an image but doesn't say what the penalty is for using the
automatic capture mode, I'm going to assume its faster).

UNFORTUNATELY!!! We can't disable the simple cam server!!! The NI class
uses a thread which is not accessible outside the object. So once we
start a simple cam server, we cannot stop it. If I'm wrong, please add
the code in to allow us to switch on the fly between manual and
automatic capture.